backend/requirements.txt
-i https://pypi.org/simple
aiohttp==3.7.4.post0; python_version >= '3.6'
aioredis==1.3.1
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.9.0; python_version >= '3.9'
argon2-cffi==20.1.0; python_version >= '3.7'
async-timeout==3.0.1; python_full_version >= '3.5.3'
asyncclick==7.1.2.3
attrs==25.3.0; python_version >= '3.8'
authlib==1.6.0; python_version >= '3.9'
cachetools==5.5.2; python_version >= '3.7'
certifi==2025.6.15; python_version >= '3.7'
cffi==1.17.1; platform_python_implementation != 'PyPy'
chardet==4.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
charset-normalizer==3.4.2; python_version >= '3.7'
click==8.2.1; python_version >= '3.10'
cloudpickle==3.1.1
colored==1.4.4; python_version >= '3.9'
colorful==0.5.6
cryptography==3.4.8; python_version >= '3.6'
docstring-parser==0.16; python_version >= '3.6' and python_version < '4.0'
environs==9.3.5; python_version >= '3.6'
fastapi==0.115.13; python_version >= '3.8'
google-adk==1.3.0; python_version >= '3.9'
google-api-core[grpc]==2.25.1; python_version >= '3.7'
google-api-python-client==2.172.0; python_version >= '3.7'
google-auth==2.40.3; python_version >= '3.7'
google-auth-httplib2==0.2.0
google-cloud-aiplatform[agent-engines]==1.97.0; python_version >= '3.9'
google-cloud-appengine-logging==1.6.2; python_version >= '3.7'
google-cloud-audit-log==0.3.2; python_version >= '3.7'
google-cloud-bigquery==3.34.0; python_version >= '3.9'
google-cloud-core==2.4.3; python_version >= '3.7'
google-cloud-logging==3.12.1
google-cloud-resource-manager==1.14.2; python_version >= '3.7'
google-cloud-secret-manager==2.24.0; python_version >= '3.7'
google-cloud-speech==2.33.0; python_version >= '3.7'
google-cloud-storage==2.19.0; python_version >= '3.7'
google-cloud-trace==1.16.2
google-crc32c==1.7.1; python_version >= '3.9'
google-genai==1.20.0; python_version >= '3.9'
google-resumable-media==2.7.2; python_version >= '3.7'
googleapis-common-protos[grpc]==1.70.0; python_version >= '3.7'
graphviz==0.21; python_version >= '3.9'
greenlet==3.2.3; python_version < '3.14' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
grpc-google-iam-v1==0.14.2; python_version >= '3.7'
grpcio==1.73.0
grpcio-status==1.73.0
h11==0.16.0; python_version >= '3.8'
hiredis==3.2.1; python_version >= '3.8'
httpcore==1.0.9; python_version >= '3.8'
httplib2==0.22.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
httpx==0.28.1; python_version >= '3.8'
httpx-sse==0.4.0; python_version >= '3.8'
idna==3.10; python_version >= '3.6'
importlib-metadata==8.7.0; python_version >= '3.9'
marshmallow==4.0.0; python_version >= '3.9'
mcp==1.9.4; python_version >= '3.10'
multidict==6.5.0; python_version >= '3.9'
numpy==2.3.0; python_version >= '3.11'
opentelemetry-api==1.34.1; python_version >= '3.9'
opentelemetry-exporter-gcp-trace==1.9.0; python_version >= '3.7'
opentelemetry-resourcedetector-gcp==1.9.0a0; python_version >= '3.7'
opentelemetry-sdk==1.34.1; python_version >= '3.9'
opentelemetry-semantic-conventions==0.55b1; python_version >= '3.9'
packaging==25.0; python_version >= '3.8'
prettyprinter==0.18.0
propcache==0.3.2; python_version >= '3.9'
proto-plus==1.26.1; python_version >= '3.7'
protobuf==6.31.1; python_version >= '3.9'
pyasn1==0.6.1; python_version >= '3.8'
pyasn1-modules==0.4.2; python_version >= '3.8'
pycparser==2.22; python_version >= '3.8'
pydantic==2.11.7; python_version >= '3.9'
pydantic-core==2.33.2; python_version >= '3.9'
pydantic-settings==2.9.1; python_version >= '3.9'
pygments==2.19.1; python_version >= '3.8'
pyparsing==3.2.3; python_version >= '3.1'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.1.0; python_version >= '3.9'
python-multipart==0.0.20; python_version >= '3.8'
pyyaml==6.0.2; python_version >= '3.8'
requests==2.32.4; python_version >= '3.8'
rsa==4.9.1; python_version >= '3.6' and python_version < '4'
shapely==2.1.1; python_version >= '3.10'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.1; python_version >= '3.7'
sqlalchemy==2.0.41; python_version >= '3.7'
sse-starlette==2.3.6; python_version >= '3.9'
starlette==0.46.2; python_version >= '3.9'
typing-extensions==4.14.0; python_version >= '3.9'
typing-inspection==0.4.1; python_version >= '3.9'
tzlocal==5.3.1; python_version >= '3.9'
uritemplate==4.2.0; python_version >= '3.9'
urllib3==2.4.0; python_version >= '3.9'
uvicore==0.1.14; python_version >= '3.7' and python_version < '4.0'
uvicorn==0.34.3; python_version >= '3.9'
websockets==15.0.1; python_version >= '3.9'
yarl==1.20.1; python_version >= '3.9'
zipp==3.23.0; python_version >= '3.9'
